RUMM-156 Move CheckLicensesPlugin into its own package
