+++ buildSrc/src/main/kotlin/com/datadog/gradle/plugin/checklicenses/ThirdPartyLicensesExtension.kt
@@ -4,7 +4,7 @@
  * Copyright 2016-2019 Datadog, Inc.
  */
 
-package com.datadog.gradle.plugin
+package com.datadog.gradle.plugin.checklicenses
 
 import java.io.File
