+++ buildSrc/src/main/kotlin/com/datadog/gradle/plugin/checklicenses/ThirdPartyLicensesPlugin.kt
@@ -4,7 +4,7 @@
  * Copyright 2016-2019 Datadog, Inc.
  */
 
-package com.datadog.gradle.plugin
+package com.datadog.gradle.plugin.checklicenses
 
 import com.android.build.gradle.internal.tasks.factory.dependsOn
 import java.io.File
@@ -16,7 +16,9 @@ class ThirdPartyLicensesPlugin : Plugin<Project> {
     override fun apply(target: Project) {
         val extension = target.extensions
             .create(EXT_NAME, ThirdPartyLicensesExtension::class.java)
-        extension.csvFile = File(target.rootDir, ThirdPartyLicensesExtension.DEFAULT_TP_LICENCE_FILENAME)
+        extension.csvFile = File(target.rootDir,
+            ThirdPartyLicensesExtension.DEFAULT_TP_LICENCE_FILENAME
+        )
 
         val updateTask = target.tasks
             .create(TASK_UPDATE_NAME, UpdateThirdPartyLicensesTask::class.java)
